+++ crates/sui-indexer-alt-graphql/schema.graphql
@@ -639,7 +639,7 @@ interface IObject {
 """
 A call to either an entry or a public Move function.
 """
-type MoveCallTransaction {
+type MoveCallCommand {
 	"""
 	Placeholder field
 	"""
@@ -937,22 +937,11 @@ type PageInfo {
 }
 
 """
-A single transaction, or command, in the programmable transaction block.
+A single command in the programmable transaction.
 """
-union ProgrammableTransaction = MoveCallTransaction
+union ProgrammableCommand = MoveCallCommand
 
-type ProgrammableTransactionBlock {
-	"""
-	Input objects or primitive values.
-	"""
-	inputs(first: Int, after: String, last: Int, before: String): TransactionInputConnection!
-	"""
-	The transaction commands, executed sequentially.
-	"""
-	transactions(first: Int, after: String, last: Int, before: String): ProgrammableTransactionConnection!
-}
-
-type ProgrammableTransactionConnection {
+type ProgrammableCommandConnection {
 	"""
 	Information to aid in pagination.
 	"""
@@ -960,25 +949,36 @@ type ProgrammableTransactionConnection {
 	"""
 	A list of edges.
 	"""
-	edges: [ProgrammableTransactionEdge!]!
+	edges: [ProgrammableCommandEdge!]!
 	"""
 	A list of nodes.
 	"""
-	nodes: [ProgrammableTransaction!]!
+	nodes: [ProgrammableCommand!]!
 }
 
 """
 An edge in a connection.
 """
-type ProgrammableTransactionEdge {
+type ProgrammableCommandEdge {
 	"""
 	The item at the end of the edge
 	"""
-	node: ProgrammableTransaction!
+	node: ProgrammableCommand!
 	"""
 	A cursor for use in pagination
 	"""
 	cursor: String!
+}
+
+type ProgrammableTransaction {
+	"""
+	Input objects or primitive values.
+	"""
+	inputs(first: Int, after: String, last: Int, before: String): TransactionInputConnection!
+	"""
+	The transaction commands, executed sequentially.
+	"""
+	commands(first: Int, after: String, last: Int, before: String): ProgrammableCommandConnection!
 }
 
 """
@@ -1565,7 +1565,7 @@ type TransactionInputEdge {
 """
 Different types of transactions that can be executed on the Sui network.
 """
-union TransactionKind = GenesisTransaction | ConsensusCommitPrologueTransaction | ChangeEpochTransaction | RandomnessStateUpdateTransaction | AuthenticatorStateUpdateTransaction | ProgrammableTransactionBlock
+union TransactionKind = GenesisTransaction | ConsensusCommitPrologueTransaction | ChangeEpochTransaction | RandomnessStateUpdateTransaction | AuthenticatorStateUpdateTransaction | ProgrammableTransaction
 
 """
 An unsigned integer that can hold values up to 2^53 - 1. This can be treated similarly to `Int`, but it is guaranteed to be non-negative, and it may be larger than 2^32 - 1.
