+++ crates/sui-indexer-alt-graphql/src/api/types/transaction_kind/programmable/commands/move_call.rs
@@ -1,0 +1,13 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use async_graphql::*;
+
+// TODO(DVX-1373): Implement MoveCallCommand
+/// A call to either an entry or a public Move function.
+#[derive(SimpleObject, Clone)]
+pub struct MoveCallCommand {
+    /// Placeholder field
+    #[graphql(name = "_")]
+    pub dummy: Option<bool>,
+}
